project/requirements.txt
opencv_python==4.3.0.36
torchvision==0.8.0
numpy==1.20.3
tensorboard_logger==0.1.0
tensorflow_gpu==2.3.0
torch==1.7.0
scipy==1.4.1
Pillow==8.3.2
tensorflow==2.6.0
